Began header navbar.
