+++ src/components/Content/index.js
@@ -1,4 +1,4 @@
 //~~~~~~~~~~~~~~~
 //~~~ EXPORTS ~~~
 //~~~~~~~~~~~~~~~
-export { default as App } from './App';
+export { default as Content } from './Content';
