+++ src/components/Footer/Footer.js
@@ -3,6 +3,11 @@
 //~~~~~~~~~~~~~~~
 import React from 'react';
 
+//--- Bootstrap ---
+import Container from 'react-bootstrap/Container';
+
+//--- Components ---
+
 //~~~~~~~~~~~~~~~~~
 //~~~ COMPONENT ~~~
 //~~~~~~~~~~~~~~~~~
