--- src/components/Header.js
@@ -1,23 +1,0 @@
-//~~~~~~~~~~~~~~~
-//~~~ IMPORTS ~~~
-//~~~~~~~~~~~~~~~
-import React from 'react';
-
-//~~~~~~~~~~~~~~~~~
-//~~~ COMPONENT ~~~
-//~~~~~~~~~~~~~~~~~
-const Header = () => {
-    //--- JSX ---
-    return (
-        <div id='header'>
-            <h1>DEFCON BUG TRACKER</h1>
-            <h3>Insert NAV BAR</h3>
-            <h3>LOGIN/LOGOUT/REGISTER</h3>
-        </div>
-    );
-}
-
-//~~~~~~~~~~~~~~~
-//~~~ EXPORTS ~~~
-//~~~~~~~~~~~~~~~
-export default Header;
