+++ src/components/Header/index.js
@@ -1,0 +1,5 @@
+//~~~~~~~~~~~~~~~
+//~~~ EXPORTS ~~~
+//~~~~~~~~~~~~~~~
+export { default as Header } from './Header';
+export { default as HeaderNav } from './HeaderNav';
