+++ src/index.js
@@ -4,6 +4,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
+import 'bootstrap/dist/css/bootstrap.min.css';
+
 import { App } from './components';
 
 ReactDOM.render(
